\name{example_map}
\alias{example_map}
\docType{data}
\title{
Example genetic maps
}
\description{
Example genetic maps including the physical position and the genetic distance of single nucleotide polymorphisms (SNPs).
}
\usage{data("example_map")}
\format{
  A data frame with 2,000 haploid genomes of the physical position, named \code{SNP}, and the genetic distance of these SNP, named \code{gd}.
}
\examples{
data(example_map)
}
\keyword{datasets}
