\name{example_painting_p1}
\alias{example_painting_p1}
\docType{data}
\title{
Example painting data for population 1.
}
\description{
Example painting data (average ancestry probabilities) of a chromosome for population 1, including 2000 genomes (observations) and 200 single nucleotide polymorphisms (SNPs).
}
\usage{data("example_painting_p1")}
\format{
  A data frame with 1500 haploid genomes (observations) on 200 SNPs (variables).
}
\examples{
data(example_painting_p1)
}
\keyword{datasets}
