% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{LDAandLDAS-package}
\alias{LDAandLDAS}
\alias{LDAandLDAS-package}
\title{Linkage Disequilibrium of Ancestry (LDA) and LDA Score (LDAS)}
\description{
This is the software for Linkage disequilibrium of ancestry (LDA) and LDA score (LDAS) which is proposed by
the papar Genetic risk for Multiple Sclerosis originated in Pastoralist Steppe populations,
Barrie W, Yang Y, Attfield K E, et al (2022).

LDA quantifies the correlations between the ancestry of two SNPs,
measuring the proportion of individuals who have experienced a recombination
leading to a change in ancestry, relative to the genome-wide baseline.

LDA score is the total amount of genome in LDA with each SNP
(measured in recombination map distance), which is useful for detecting the
signal of “recombinant favouring selection”.

The codes for LDA and LDAS are hosted at \url{https://github.com/YaolingYang/LDAandLDAscore}.
}
\references{
Barrie W, Yang Y, Attfield K E, et al. Genetic risk for Multiple Sclerosis originated in Pastoralist Steppe populations. bioRxiv (2022).
}
\author{
\strong{Maintainer}: Yaoling Yang \email{yaoling.yang@bristol.ac.uk} (\href{https://orcid.org/0000-0003-4905-8097}{ORCID})

Authors:
\itemize{
  \item Daniel Lawson \email{Dan.Lawson@bristol.ac.uk} (\href{https://orcid.org/0000-0002-5311-6213}{ORCID})
}

}
